+++ lib/rpc.ex
@@ -5,12 +5,15 @@ defmodule Soroban.RPC do
 
   alias Soroban.RPC.{
     GetEvents,
+    GetFeeStats,
     GetHealth,
     GetLatestLedger,
     GetLedgerEntries,
     GetLedgerEntriesResponse,
     GetNetwork,
     GetTransaction,
+    GetTransactions,
+    GetVersionInfo,
     SendTransaction,
     SimulateTransaction
   }
@@ -36,11 +39,14 @@ defmodule Soroban.RPC do
     as: :request
 
   defdelegate get_transaction(server, hash), to: GetTransaction, as: :request
+  defdelegate get_transactions(server, payload), to: GetTransactions, as: :request
   defdelegate get_ledger_entries(server, keys), to: GetLedgerEntries, as: :request
   defdelegate get_events(server, payload), to: GetEvents, as: :request
   defdelegate get_health(server), to: GetHealth, as: :request
   defdelegate get_latest_ledger(server), to: GetLatestLedger, as: :request
   defdelegate get_network(server), to: GetNetwork, as: :request
+  defdelegate get_fee_stats(server), to: GetFeeStats, as: :request
+  defdelegate get_version_info(server), to: GetVersionInfo, as: :request
 
   defp encode_xdr(%LedgerKey{} = ledger_key) do
     ledger_key